moved-logic-to-shared-utils
